requirements.txt
# Core Robot Framework
robotframework==7.0
robotframework-seleniumlibrary==6.2.0

# Data handling
openpyxl==3.1.2
pandas==2.2.2
robotframework-datadriver==1.8.0

# Optional utilities
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.6


